requirements.txt
# LangGraph Invoice Processing Agent Requirements
# Python 3.10+

# Core LangGraph and LangChain
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Data Processing
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# OCR and Document Processing
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0

# HTTP and API
requests>=2.31.0
httpx>=0.25.0

# Utilities
python-dateutil>=2.8.2
structlog>=23.2.0
uuid>=1.30

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# MCP Protocol (Mock implementation)
# Note: In production, use actual MCP SDK
pydantic-settings>=2.0.0

# Tool Selection
openai>=1.0.0  # For LLM-based tool selection in Bigtool
